fix: update critical flow tests file structure [WPB-18713]
